requirements.txt
amqp==5.3.1
asgiref==3.9.1
async-timeout==5.0.1
billiard==4.2.1
celery==5.5.3
certifi==2025.8.3
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.5
Django==5.2.4
django-celery-beat==2.8.1
django-redis==6.0.0
django-timezone-field==7.1
djangorestframework==3.16.0
idna==3.10
kombu==5.5.4
packaging==25.0
prompt_toolkit==3.0.51
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-decouple==3.8
redis==6.2.0
requests==2.32.4
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.13
